week6/requirements.txt
pymongo
pika
jsonrpclib-pelix
pylint
requests
redis
lxml